refactor: add a single global object for storing text
