+++ .gitignore
@@ -26,6 +26,7 @@ share/python-wheels/
 *.egg
 MANIFEST
 **/*.DS_Store
+!/apps/react-ui/client/src/lib
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
